refactor: remove statics from Renderable
