+++ src/main/java/org/runejs/client/cache/media/gameInterface/GameInterface.java
@@ -2293,6 +2293,50 @@ ChatBox.tradeMode
         interfaceTypefaceCache.clear();
     }
 
+    public static boolean handleSequences(int arg1) {
+        if(!decodeGameInterface(arg1))
+            return false;
+        GameInterface[] gameInterfaces = cachedInterfaces[arg1];
+        boolean bool = false;
+        for(int i = 0; gameInterfaces.length > i; i++) {
+            GameInterface gameInterface = gameInterfaces[i];
+            if(gameInterface != null && gameInterface.type == GameInterfaceType.MODEL) {
+                if(gameInterface.animation != -1 || gameInterface.alternateAnimation != -1) {
+                    boolean bool_0_ = checkForAlternateAction(gameInterface);
+                    int i_1_;
+                    if(bool_0_)
+                        i_1_ = gameInterface.alternateAnimation;
+                    else
+                        i_1_ = gameInterface.animation;
+                    if(i_1_ != -1) {
+                        AnimationSequence animationSequence = AnimationSequence.getAnimationSequence(i_1_);
+                        gameInterface.remainingAnimationTime += MovedStatics.anInt199;
+                        while(animationSequence.frameLengths[gameInterface.animationFrame] < gameInterface.remainingAnimationTime) {
+                            bool = true;
+                            gameInterface.remainingAnimationTime -= animationSequence.frameLengths[gameInterface.animationFrame];
+                            gameInterface.animationFrame++;
+                            if(gameInterface.animationFrame >= animationSequence.frameIds.length) {
+                                gameInterface.animationFrame -= animationSequence.frameStep;
+                                if(gameInterface.animationFrame < 0 || animationSequence.frameIds.length <= gameInterface.animationFrame)
+                                    gameInterface.animationFrame = 0;
+                            }
+                        }
+                    }
+                }
+                if(gameInterface.rotationSpeed != 0) {
+                    bool = true;
+                    int i_2_ = gameInterface.rotationSpeed >> 16;
+                    int i_3_ = gameInterface.rotationSpeed << 16 >> 16;
+                    i_2_ *= MovedStatics.anInt199;
+                    gameInterface.rotationX = 0x7ff & i_2_ + gameInterface.rotationX;
+                    i_3_ *= MovedStatics.anInt199;
+                    gameInterface.rotationZ = 0x7ff & gameInterface.rotationZ + i_3_;
+                }
+            }
+        }
+        return bool;
+    }
+
     public void swapItems(int arg0, boolean arg1, int arg2) {
         int i = items[arg2];
         items[arg2] = items[arg0];
